feat: ReactSolutionView
