+++ src/routes/PublicRoutes.tsx
@@ -5,6 +5,7 @@ import AboutView from '../views/About';
 import DevView from '../views/Dev';
 import HomeView from '../views/Home';
 import MuiSolutionView from '../views/MuiSolution';
+import ReactSolutionView from '../views/ReactSolution';
 
 /**
  * List of routes available for anonymous users
@@ -17,6 +18,7 @@ const PublicRoutes = () => {
       <Routes>
         <Route path="/" element={<HomeView />} />
         <Route path="mui" element={<MuiSolutionView />} />
+        <Route path="react" element={<ReactSolutionView />} />
         <Route path="about" element={<AboutView />} />
         {process.env.REACT_APP_DEBUG === 'true' && <Route path="dev" element={<DevView />} />}
         <Route path="*" element={<NotFoundView />} />
