+++ src/views/ReactSolution/index.tsx
@@ -1,0 +1,3 @@
+import ReactSolutionView from './ReactSolutionView';
+
+export default ReactSolutionView;
